+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Introduce the `cors`, `parallel` middlewares.
 - Introduce `set` options for `request_transformer` and `response_transformer` middlewares.
+- Introduce `nacos` config provider.
 
 ### **Bug Fixes**